organize into vendor submod
