+++ lib/src/music/spotify/mod.rs
@@ -1,7 +1,6 @@
 pub mod client;
+pub mod vendor;
 
-use entity::{album, artist};
-use sea_orm::{prelude::Date, ActiveValue, NotSet};
 use serde::{Deserialize, Serialize};
 use surf::Body;
 
@@ -11,111 +10,28 @@ pub struct SpotifyError {
     pub message: String,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Track {
-    pub name: String,
-    pub album: Album,
-    external_urls: ExternalUrls,
-}
-
-impl Into<entity::track::Entity> for Track {
-    fn into(self) -> entity::prelude::Track {
-        entity::prelude::Track {}
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Artist {
-    pub name: String,
-    external_urls: ExternalUrls,
-}
-
-impl Artist {
-    pub fn model(&self) -> artist::ActiveModel {
-        artist::ActiveModel {
-            id: NotSet,
-            name: ActiveValue::set(self.name.clone()),
-            created_at: NotSet,
-            updated_at: NotSet,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Album {
-    images: Vec<AlbumImage>,
-    pub name: String,
-    release_date: String,
-    album_type: String,
-    external_urls: ExternalUrls,
-    pub artists: Vec<Artist>,
-}
-
-impl Album {
-    pub fn model(&self) -> album::ActiveModel {
-        let release_date = Date::parse_from_str(&self.release_date, "%Y-%m-%d")
-            .expect("Failed to parse release date");
-        album::ActiveModel {
-            id: NotSet,
-            title: ActiveValue::set(self.name.clone()),
-            release_date: ActiveValue::set(release_date),
-            created_at: NotSet,
-            updated_at: NotSet,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct AlbumImage {
-    url: String,
-    width: u32,
-    height: u32,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct ExternalUrls {
-    spotify: String,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct RecentTrack {
-    pub track: Track,
-    pub played_at: String,
-}
-
-impl Track {
-    pub fn model(&self) -> entity::track::ActiveModel {
-        entity::track::ActiveModel {
-            id: NotSet,
-            title: ActiveValue::set(self.name.clone()),
-            created_at: NotSet,
-            updated_at: NotSet,
-        }
-    }
-}
-
 pub trait RecentTrackExt {
-    fn artists(&self) -> Vec<Artist>;
-    fn albums(&self) -> Vec<Album>;
+    fn artists(&self) -> Vec<vendor::Artist>;
+    fn albums(&self) -> Vec<vendor::Album>;
 }
 
-impl RecentTrackExt for Vec<RecentTrack> {
+impl RecentTrackExt for Vec<vendor::RecentTrack> {
     /// Gets all the artists from the recent tracks
-    fn artists(&self) -> Vec<Artist> {
+    fn artists(&self) -> Vec<vendor::Artist> {
         self.iter()
             .map(|track| track.track.album.artists.clone())
             .flatten()
             .collect()
     }
     /// Gets all the albums from the recent tracks
-    fn albums(&self) -> Vec<Album> {
+    fn albums(&self) -> Vec<vendor::Album> {
         self.iter().map(|track| track.track.album.clone()).collect()
     }
 }
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct RecentTracksResponse {
-    pub items: Option<Vec<RecentTrack>>,
+    pub items: Option<Vec<vendor::RecentTrack>>,
     pub error: Option<SpotifyError>,
 }
